default headers are up and running! Deployment almost ready, time to switch repos

+++ WindowCreation/include/embedded_resources.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <cstddef>
 
 /* EMBEDDED RESOURCES HEADER FILE
 -------------------------------------------------------------------------------------------------------
@@ -52,4 +51,4 @@ enum class BLOB_ID
 const void* getBlobFromId(BLOB_ID id) noexcept;
 
 // Returns the size in bytes of the blob data.
-size_t getBlobSizeFromId(BLOB_ID id) noexcept;
+unsigned long long getBlobSizeFromId(BLOB_ID id) noexcept;
